move frontend utils functions to @frontend/utils (#107)
