+++ src/frontend/components/devPanel/renderStatus.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import * as Types from "@data-types/index";
 
-import { getStatusIcon, getStatusMessage } from "@utils/index";
+import { getStatusIcon, getStatusMessage } from "@frontend/utils/index";
 
 const RenderStatus = ({ status }: { status: Types.Context.Status }) => (
   <div className="block">
